pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=5.0.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=4.1.2",
    "fastapi>=0.124.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.38.0",
]
